+++ packages/extension-base/src/services/earning-service/handlers/liquid-staking/stella-swap.ts
@@ -3,12 +3,12 @@
 
 import { TransactionError } from '@subwallet/extension-base/background/errors/TransactionError';
 import { BasicTxErrorType, ChainType, ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
-import { getERC20Contract } from '@subwallet/extension-base/koni/api/tokens/evm/web3';
+import { getERC20Contract, getERC20SpendingApprovalTx } from '@subwallet/extension-base/koni/api/contract-handler/evm/web3';
 import KoniState from '@subwallet/extension-base/koni/background/handlers/State';
 import { _EvmApi } from '@subwallet/extension-base/services/chain-service/types';
 import { _getAssetDecimals, _getContractAddressOfToken } from '@subwallet/extension-base/services/chain-service/utils';
 import { calculateGasFeeParams } from '@subwallet/extension-base/services/fee-service/utils';
-import { BaseYieldStepDetail, EarningStatus, HandleYieldStepData, LiquidYieldPoolInfo, OptimalYieldPath, OptimalYieldPathParams, SubmitYieldJoinData, TokenApproveData, TransactionData, UnstakingInfo, UnstakingStatus, YieldPoolMethodInfo, YieldPositionInfo, YieldStepType, YieldTokenBaseInfo } from '@subwallet/extension-base/types';
+import { BaseYieldStepDetail, EarningStatus, HandleYieldStepData, LiquidYieldPoolInfo, OptimalYieldPath, OptimalYieldPathParams, SubmitYieldJoinData, TokenSpendingApprovalParams, TransactionData, UnstakingInfo, UnstakingStatus, YieldPoolMethodInfo, YieldPositionInfo, YieldStepType, YieldTokenBaseInfo } from '@subwallet/extension-base/types';
 import { TransactionConfig } from 'web3-core';
 import { Contract } from 'web3-eth-contract';
 
@@ -37,8 +37,6 @@ interface StellaswapUnbonding {
   unbonded: string,
   waiting: string
 }
-
-const MAX_INT = '115792089237316195423570985008687907853269984665640564039457584007913129639935';
 
 export default class StellaSwapLiquidStakingPoolHandler extends BaseLiquidStakingPoolHandler {
   public slug: string;
@@ -316,34 +314,18 @@ export default class StellaSwapLiquidStakingPoolHandler extends BaseLiquidStakin
     const derivativeTokenInfo = this.state.getAssetBySlug(this.derivativeAssets[0]);
     const derivativeTokenContractAddress = _getContractAddressOfToken(derivativeTokenInfo);
     const evmApi = this.evmApi;
-    const inputTokenContract = getERC20Contract(_getContractAddressOfToken(inputTokenInfo), evmApi);
+    const transactionObject = await getERC20SpendingApprovalTx(derivativeTokenContractAddress, address, _getContractAddressOfToken(inputTokenInfo), evmApi);
 
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-call,@typescript-eslint/no-unsafe-member-access,@typescript-eslint/no-unsafe-assignment
-    const approveCall = inputTokenContract.methods.approve(derivativeTokenContractAddress, MAX_INT); // TODO: need test
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-call,@typescript-eslint/no-unsafe-member-access,@typescript-eslint/no-unsafe-assignment
-    const approveEncodedCall = approveCall.encodeABI() as string;
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-call,@typescript-eslint/no-unsafe-member-access
-    const gasLimit = await approveCall.estimateGas({ from: address }) as number;
-    const priority = await calculateGasFeeParams(evmApi, this.chain);
-
-    const transactionObject = {
-      from: address,
-      to: _getContractAddressOfToken(inputTokenInfo),
-      data: approveEncodedCall,
-      gas: gasLimit,
-      gasPrice: priority.gasPrice,
-      maxFeePerGas: priority.maxFeePerGas?.toString(),
-      maxPriorityFeePerGas: priority.maxPriorityFeePerGas?.toString()
-    } as TransactionConfig;
-
-    const _data: TokenApproveData = {
-      inputTokenSlug: inputTokenSlug,
-      spenderTokenSlug: this.derivativeAssets[0]
+    const _data: TokenSpendingApprovalParams = {
+      contractAddress: inputTokenSlug,
+      spenderAddress: this.derivativeAssets[0],
+      owner: address,
+      chain: this.chain
     };
 
     return Promise.resolve({
       txChain: this.chain,
-      extrinsicType: ExtrinsicType.TOKEN_APPROVE,
+      extrinsicType: ExtrinsicType.TOKEN_SPENDING_APPROVAL,
       extrinsic: transactionObject,
       txData: _data,
       transferNativeAmount: '0',