+++ packages/extension-koni-ui/src/Popup/Confirmations/variants/Transaction/variants/SendNft.tsx
@@ -44,9 +44,9 @@ const Component: React.FC<Props> = (props: Props) => {
       </MetaInfo>
       <MetaInfo hasBackgroundWrapper={true}>
         {
-          (data.nftItemName || data.nftItem) && (
+          !!(data.nftItemName || data.nftItem) && (
             <MetaInfo.Default label={t('NFT')}>
-              {data.nftItemName || data.nftItem.name || `${data.nftItem.collectionId}_${data.nftItem.id}`}
+              {data.nftItemName || data.nftItem.name || data.nftItem.id}
             </MetaInfo.Default>
           )
         }