+++ packages/extension-koni/src/helper/ActionHandler.ts
@@ -109,10 +109,10 @@ export class ActionHandler {
       // Set timeout to sleep
       if (Object.keys(this.connectionMap).length === 0) {
         console.debug('Every port is disconnected, set timeout to sleep');
+        this.isActive = false;
         this.sleepTimeout && clearTimeout(this.sleepTimeout);
         this.sleepTimeout = setTimeout(() => {
           // Reset active status
-          this.isActive = false;
           this.waitActiveHandler = createPromiseHandler<boolean>();
           this.mainHandler && this.mainHandler.state.sleep().catch(console.error);
         }, SLEEP_TIMEOUT);